requirements.txt
pyyaml
requests >= 2.2.1